requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.13

# AI and ML
# OpenAI - Latest stable version for GPT integration
openai==1.6.1
# LangChain for AI workflows
langchain==0.0.350
langchain-openai==0.0.2
# Pinecone for vector search and embeddings
pinecone-client==2.2.4

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Async and concurrency
asyncio-mqtt==0.13.0
websockets==12.0
aiofiles==23.2.1
tenacity==8.2.3

# HTTP clients - Compatible versions
python-multipart==0.0.6
aiohttp==3.9.1

# Logging and monitoring
structlog==23.2.0

# Development tools
black==23.11.0
flake8==6.1.0

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
serpapi==0.1.5

gunicorn
